backend/screener_server/requirements.txt
fastapi
uvicorn
pydantic
selenium
undetected-chromedriver
requests
google-genai
setuptools
python-dotenv